moved RangerHdfsPlugin
